Progress on #10: Handling variable yScale, ickily.
Not sure how best to do this. So the problem is that closure-scope
variables depend on what selection/data is passed in, which leaves a
"residue" that has to be reset every time. What's the "right" pattern
here? Using internal getter/setters to access variable properties?

I mean, only the height is *really* changing, right? Everything else
(scale and in turn axis) just accommodates that?

I'm not sure there's a fundamentally better way to do this, except
caching, which wouldn't scale if I wanted to do a whole bunch of
different heights... which I don't. But it can certainly be tidied up,
refactored the slightest little superficial bit.